Make building tests optional
As libuseful is a library it is useful to disable building tests to speed
up the build times of applications that depend on it.  To do this, set the
LIBUSEFUL_BUILD_TESTS cmake option appropriately.  By default it is on to
maintain the current default.

The option may be set in the CMakeLists.txt of the dependent application:

	SET(LIBUSEFUL_BUILD_TESTS OFF)

It can also be set on the cmake command line:

	cmake -DLIBUSEFUL_BUILD_TESTS=OFF ...